implement state in addRecipe to track the form data
This commit moves the state out of addrecipeingredients and into
addrecipe. This is so that when we submit the form an ajax request can
read the state of addrecipe and send the state to the server. The
handleSubmit function is not fully implemented yet, but it will be used
to actually send the data.