Change clientID and Secret to use env variables
this change was necessary to make things work with Heroku. Heroku gets
files directly from github, but we can't check in files with keys in
them. The way around this is to use config vars in heroku. These set up
environmental variables that you can set to be equal to they key values.
See https://devcenter.heroku.com/articles/config-vars. Node will reach
into the environment using process.env and look for these environmental
variables and fill them in in this file.